Add basic rotation stuff
Tried a Hough transform but it didn't work so well. Should probably try
this again though because apparently it should work for this kind of
thing. Other good option could be PCA. In any case, for now I've just
done a silly bucketing thing with the angles. It seems to do moderately
well.