Extract the truncated, rotated crossword
Turns out a really simple method worked much better than the magic: do
edge detection, fill from the four corners of the image, take the
average white pixel position (should be in the crossword), fill from
there allowing arbitrary colour changes but using the previously-filled
area as a mask. The component filled with that list fill is the
crossword. Then we can do edge detection and Hough on the mask to get
the angle.