Detect square colour
Also generate a grid based on the colours of each square. To get the
colours we just get a point in the square (by assuming it's a perfect
grid and that our grid size is correct), take a rectangle around that
point, and see whether more pixels are light or dark.