Initial caching implementation
This adds an index to the ggc, and uses that when hashing
image data to ensure that we re-get when the ggc changes. This isn't
great, because then we unnecessarily store a bunch of images we'll never
use again (because typically each image will only ever be used with one
or two different ggcs). Instead, we should just clear our the cache
whenever the ggc changes (there's a TODO for that).