Basic setup for deciding when to load progress
Each time we show the cw page we sent a "cache key". If the cache key is
the same as last time, the frontend should try to load saved progress.

The cache key is a hash of image data and semantic data. This way if the
backend changes and causes the semantic data to change, or if the user
happens to upload a different crossword with the same semantic data, we
won't mistakenly try to load progress.