Save progress to local storage
Turns out localStorage is what you're meant to use, not cookies. Anyway,
as intended last time, on load we check the keys and if they match we
attempt to load, if they don't then we forget the old data. Every time
something is input we save everything.

Just in case, we also add a button to clear saved data (in case
something weird is happening).